requirements.txt
requests>=2.28.0
python-whois>=0.7.3
colorama>=0.4.4
beautifulsoup4>=4.11.0
python-nmap>=0.7.1
Pillow>=9.0.0
